requirements.txt
tensorflow-cpu==2.15.0
numpy
Pillow
h5py
gradio
scipy
